# App/Back/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Fithub API
  description: API para la plataforma de entrenamiento Fitter
  version: 1.0.0-SNAPSHOT
servers:
  - url: http://localhost:8080/api
paths:
  /usuarios:
    get:
      summary: Obtener todos los usuarios
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      summary: Crear nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '201':
          description: Usuario creado
  /usuarios/{id}:
    get:
      summary: Obtener usuario por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
    put:
      summary: Actualizar usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '200':
          description: Usuario actualizado
    delete:
      summary: Eliminar usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado
...
  /horarios/usuario/{id}:
    get:
      summary: Horarios asignados a un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de horarios del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Horario'
...
  /compras/usuario/{id}:
    get:
      summary: Ver compras hechas por un usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de compras del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Compra'
...
components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        correo:
          type: string
        contraseña:
          type: string
        rolId:
          type: integer
      required:
        - nombre
        - correo
        - contraseña
...
    Horario:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        dia:
          type: string
        hora:
          type: string
        planId:
          type: integer
      required:
        - usuarioId
        - dia
        - hora
        - planId
    Compra:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        fecha:
          type: string
      required:
        - usuarioId
        - fecha
...